Move next two methods to util.ts
